backend/requirements.txt
Django>=4.3
djangorestframework>=3.17.1,<4
djangorestframework-simplejwt>=2.0
django-cors-headers>=4.0
